+++ mona/include/monapi/messages.h
@@ -57,7 +57,7 @@ extern monapi_cmemoryinfo* monapi_call_file_decompress_st5_file(const char* file
 extern int monapi_call_process_execute_file(const char* command_line, MONAPI_BOOL prompt);
 extern int monapi_call_process_execute_file_get_tid(const char* command_line, MONAPI_BOOL prompt, uint32_t* tid, uint32_t stdout_id, uint32_t stdin_id);
 extern int monapi_call_change_drive(int drive, MONAPI_BOOL prompt);
-extern uint32_t monapi_file_open(const char* file, MONAPI_BOOL create);
+extern intptr_t monapi_file_open(const char* file, MONAPI_BOOL create);
 extern uint32_t monapi_file_write(uint32_t fileID, monapi_cmemoryinfo* mem, uint32_t size);
 extern monapi_cmemoryinfo* monapi_file_read(uint32_t fileID, uint32_t size);
 extern uint32_t monapi_file_seek(uint32_t fileID, uint32_t offset, uint32_t origin);
